split About Tab from profile editor popup
