+++ client/src/components/Styles/profileEditor.css
@@ -13,7 +13,7 @@ textarea {
   width: 90vw;
   max-width: 1120px;
   height: calc(90vh - 200px);
-  /* padding: 50px; */
+  padding: 50px;
   border: none;
   border-radius: 10px;
   color: var(--main-text);
